@@ -2,75 +2,77 @@ package cafe.model.entity;
 
 import java.io.Serializable;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.NamedQuery;
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.NamedQuery;
 
-@XmlRootElement
 @Entity
 @NamedQuery(name = "findAllCoffees", query = "SELECT o FROM Coffee o")
 public class Coffee implements Serializable {
 
-  private static final long serialVersionUID = 1L;
+	private static final long serialVersionUID = 1L;
 
-  @Id @GeneratedValue private Long id;
-  protected String name;
-  protected Double price;
+	@Id
+	@GeneratedValue
+	private Long id;
+	private String name;
+	private Double price;
 
-  public Coffee() {}
+	public Coffee() {
+		// Nothing to do.
+	}
 
-  public Coffee(String name, Double price) {
-    this.name = name;
-    this.price = price;
-  }
+	public Coffee(String name, Double price) {
+		this.name = name;
+		this.price = price;
+	}
 
-  public String getName() {
-    return name;
-  }
+	public String getName() {
+		return name;
+	}
 
-  public void setName(String name) {
-    this.name = name;
-  }
+	public void setName(String name) {
+		this.name = name;
+	}
 
-  public Double getPrice() {
-    return price;
-  }
+	public Double getPrice() {
+		return price;
+	}
 
-  public void setPrice(Double price) {
-    this.price = price;
-  }
+	public void setPrice(Double price) {
+		this.price = price;
+	}
 
-  public Long getId() {
-    return id;
-  }
+	public Long getId() {
+		return id;
+	}
 
-  public void setId(Long id) {
-    this.id = id;
-  }
+	public void setId(Long id) {
+		this.id = id;
+	}
 
-  @Override
-  public int hashCode() {
-    int hash = 0;
-    hash += (id != null ? id.hashCode() : 0);
-    return hash;
-  }
+	@Override
+	public int hashCode() {
+		int hash = 0;
+		hash += (id != null ? id.hashCode() : 0);
+		return hash;
+	}
 
-  @Override
-  public boolean equals(Object object) {
-    if (!(object instanceof Coffee)) {
-      return false;
-    }
-    Coffee other = (Coffee) object;
-    if ((this.id == null && other.id != null) || (this.id != null && !this.id.equals(other.id))) {
-      return false;
-    }
-    return true;
-  }
+	@Override
+	public boolean equals(Object object) {
+		if (!(object instanceof Coffee)) {
+			return false;
+		}
+		Coffee other = (Coffee) object;
+		if ((this.id == null && other.id != null) || (this.id != null && !this.id.equals(other.id))) {
+			return false;
+		}
+		return true;
+	}
 
-  @Override
-  public String toString() {
-    return "cafe.model.entity.Coffee[id=" + id + ", name=" + name + ", price=" + price + "]";
-  }
+	@Override
+	public String toString() {
+		return "cafe.model.entity.Coffee[id=" + id + ", name=" + name + ", price=" + price + "]";
+	}
 }
